refactor(renderer): extract app-shell resume cluster into use-shell-resume (R2)
The #1223 safe-boundary resume cluster moves out of app-shell.tsx into a new
use-shell-resume.ts hook, following the use-shell-connections / use-shell-chat-model
house style (options object; state + handler returned; stable identities preserved).
Pure move, zero behavior change: the two useState declarations (resumePendingSessionId,
resumeParkDescriptionBySession) and the resumeInterruptedSession handler move verbatim;
activeId/toastApi/shellCopy/uiLocale are injected as options. app-shell keeps the banner
JSX wiring (safeResumeAction=) and the sendWithAttachments guard. resumeParkToastCopy is
no longer imported into app-shell (it now lives in the hook). app-shell.tsx 1686 -> 1654
(-32 lines).

Contract re-pin (never deletes coverage): runtime-resume-routing-contract.test.ts read
the handler-shape assertions (resumeInterruptedSession / resumeLatest(sessionId) /
resumeParkToastCopy / no rejectionReasons.join) from app-shell.tsx single-file; they now
read use-shell-resume.ts where that logic lives, while the app-shell-specific assertions
(sendWithAttachments '/resume' guard, safeResumeAction= wiring) still read app-shell.tsx.
use-shell-resume.ts added to renderer-shell-source-helpers sourcePaths (Round B/E
precedent) so combined-source contracts see it.

Gates: desktop 2740/2740, ui 196/196, typecheck 0, check-dead-css clean, knip
desktop+ui exit 0, AUDIT_PORT_BASE=23900 alignment auditor exit 0 (all 10 fixtures
clean, real renderer). CDP turn-narrative branch-vs-baseline (real Electron, light+dark
1280) byte-identical (sha256 match) — proven render no-op.
