+++ apps/desktop/src/main/__tests__/renderer-shell-source-helpers.ts
@@ -36,6 +36,7 @@ const sourcePaths = [
   'use-shell-connections.ts',
   'use-shell-chat-model.ts',
   'use-shell-live-turn.ts',
+  'use-shell-resume.ts',
   'use-shell-expert-teams.ts',
   'use-shell-memory-pill.ts',
   'use-shell-layout.ts',
